# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kubernetes FileBrowser",
        "title": "Kube-FileBrowser Swagger",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/kubeapiproxy/namespace": {
            "get": {
                "description": "命名空间列表",
                "tags": [
                    "Kubernetes Api Proxy"
                ],
                "summary": "Namespace",
                "parameters": [
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/kubeapiproxy/pods": {
            "get": {
                "description": "命名空间下Pod资源列表",
                "tags": [
                    "Kubernetes Api Proxy"
                ],
                "summary": "Pods",
                "parameters": [
                    {
                        "type": "string",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "pods",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "pods",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "pods",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "field_selector",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "label_selector",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "pods",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/kubeapiproxy/terminal": {
            "get": {
                "description": "pod 中容器的终端",
                "tags": [
                    "Kubernetes Api Proxy"
                ],
                "summary": "Container terminal",
                "parameters": [
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "shell",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "shell",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "shell",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "shell",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "消息"
                }
            }
        }
    }
}